Update CreateAndPrintMyParcelTrack.php
While working on a bug for a customer that uses this plugin I noticed this bug in the error handling code. The getMessages function returns a Collection object that does not implement a Countable interface. It does however have a getItems() function that returns an array which is countable. Alternatively the Collection class also contains a getCount function which could be used.