Fix issue with fetching order with non-numeric increment ID
When using non-numeric order increment IDs, the script will load
all orders in the webshop, causing an out of memory error when
having a lot of orders.

This is fixed by always using a string, since order increment IDs
in Magento are stored in a VARCHAR table column.

To prevent possible SQL injections, a check has been added to the
script as well.