Fix: use correct way to inject search value
When using an order prefix in Magento (like `ORD00000123456` instead of just `00000123456`), the `$searchValue` is cast to an `int`.

This presumably done to avoid SQL injections, so I changed it to using the second parameter of `where()` so Magento escapes it for us and SQL injections are avoided. This way, we can use strings to look for orders with an increment ID instead of a numeric value.